data-ingest/requirements.txt
# Data Ingest Module Requirements
feedparser>=6.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dateutil>=2.8.0
lxml>=4.9.0
